reworked examples and ENABLED fixes
