+++ CMakePresets.json
@@ -13,6 +13,7 @@
         "ENABLE_ALL_OT": true,
         "ENABLE_SSE": true,
         "ENABLE_AVX": true,
+        "ENABLE_BOOST": true,
         "ENABLE_BITPOLYMUL": false,
         "ENABLE_CIRCUITS": true,
         "LIBOTE_STD_VER": "17",
@@ -43,17 +44,13 @@
       "cacheVariables": {
         "CMAKE_BUILD_TYPE": "Debug",
         "ENABLE_INSECURE_SILVER": false,
-        "ENABLE_LDPC": false,
+        "ENABLE_PPRF": true,
+        "ENABLE_SILENT_VOLE": true,
         "LIBOTE_STD_VER": "17",
-        "ENABLE_ALL_OT": true,
-        "ENABLE_KKRT": "ON",
-        "ENABLE_IKNP": "ON",
-        "ENABLE_MR": "ON",
-        "ENABLE_SIMPLESTOT": "ON",
         "ENABLE_RELIC": false,
         "ENABLE_SODIUM": true,
-        "ENABLE_BOOST": false,
-        "ENABLE_BITPOLYMUL": true,
+        "ENABLE_BOOST": true,
+        "ENABLE_BITPOLYMUL": false,
         "FETCH_AUTO": "ON",
         "ENABLE_CIRCUITS": true,
         "VERBOSE_FETCH": true,
@@ -88,7 +85,7 @@
         "ENABLE_ALL_OT": true,
         "ENABLE_RELIC": true,
         "ENABLE_SODIUM": false,
-        "ENABLE_BOOST": false,
+        "ENABLE_BOOST": true,
         "ENABLE_OPENSSL": false,
         "FETCH_AUTO": true,
         "ENABLE_CIRCUITS": true,
