+++ cmake/buildOptions.cmake.in
@@ -70,6 +70,7 @@ set(ENABLE_KKRT           @ENABLE_KKRT@)
 set(ENABLE_SILENT_VOLE    @ENABLE_SILENT_VOLE@)
 set(NO_SILVER_WARNING     @NO_SILVER_WARNING@)
 
+set(ENABLE_PPRF           @ENABLE_PPRF@)
 
 set(libOTe_boost_FOUND		${ENABLE_BOOST})
 set(libOTe_relic_FOUND		${ENABLE_RELIC})
